+++ PyEMD/__init__.py
@@ -1,0 +1,10 @@
+import logging
+
+from PyEMD.EMD import EMD
+from PyEMD.EEMD import EEMD
+
+logger = logging.getLogger('pyemd')
+logger.addHandler(logging.NullHandler())
+if len(logger.handlers)==1:
+    logging.basicConfig(level=logging.INFO)
+
